Use Postgres for listuser
Originally, I use a json file as a temporary database.
In this commit, I replace the temporary one with Postgres.
That is, I create a database with Postgres for users, and now,
when a user send listusers to the server, the server will get
data from the table in Postgres.