Add unit-test for init database
A script file, unit-test.sh is added for testing. This time,
testing for initializing database is added. First, the script run
docker compose build and up to start needed containers. After checking
the containers start successfully, wait-for-it.sh is run to wait
for postgres database. However, docker might stop the database
container and then run the second database container. If the checking
script is run immediately after the first database container is run,
the script will try to access the database when the database is
removed. Therefore, 20s is added to wait for the second database
container.